drake-simulation-server/requirements.txt
fastapi
uvicorn[standard]
python-socketio
drake
ipywidgets
